backend/pyproject.toml
[project]
name = "finally-backend"
version = "0.1.0"
description = "FinAlly trading workstation backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",
    "litellm>=1.40.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["market", "db", "routes", "llm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
]
